dumb

+++ Classes/CharacterData.h
@@ -28,8 +28,8 @@ public:
     CC_SYNTHESIZE(int, attack, Attack);
     CC_SYNTHESIZE(int, defense, Defense);
     
-    CC_SYNTHESIZE(int, speed, Speed);
-    CC_SYNTHESIZE(int, jumpHeight, JumpHeight);
+    CC_SYNTHESIZE(float, speed, Speed);
+    CC_SYNTHESIZE(float, jumpHeight, JumpHeight);
     CC_SYNTHESIZE(float, attackSpeed, AttackSpeed);
     CC_SYNTHESIZE(int, maxJumpTimes, MaxJumpTimes);
